requirements.txt
langchain==0.3.25
langchain_anthropic==0.3.14
langchain_google_genai==2.1.5
langchain_huggingface==0.2.0
langchain_openai==0.3.18
python-dotenv==1.1.0
